Remove enqueue as a callback
I think that the enqueue action should be handle by who
needs it, so it does not belongs to this project.
I'm starting to move this behaviour into a service and with
that said I think that the service should be responsable on
asking for an enqueue.
Also, I don't think that enqueue should be a private method.
Maybe we need to handle this action from an admin page.